Move identifier pytest fixtures into conftest.py
